fix(relay): apply the gh cwd fallback to the /rpc path too
The box's forwarder posts to the control plane's /rpc route, so a cloud box's
gh.pr.* is served by server.ts — not the cloud executor patched in bc561338 —
and still spawned gh in the (nonexistent) host repo. Move ghRunContext into
gh.ts, which both dispatch paths already import, and use it in all three gh
handlers on each side.
