+++ packages/relay/src/host-actions.ts
@@ -15,7 +15,7 @@
  */
 
 import { execa } from 'execa';
-import { repoSlugFromRemote, toHttpsUrl } from './git-pat.js';
+import { toHttpsUrl } from './git-pat.js';
 import { existsSync } from 'node:fs';
 import { mkdtemp, rm } from 'node:fs/promises';
 import { tmpdir } from 'node:os';
@@ -60,6 +60,7 @@ import {
   refuseCheckoutByDefault,
   refuseGhApiCall,
   refuseMergeBypass,
+  ghRunContext,
   runHostGh,
   type GhApiRpcParams,
   type GhPrRpcParams,
@@ -382,34 +383,6 @@ export async function executeCloudAction(
     stdout: '',
     stderr: `host executor for '${action.method}' is not yet supported for cloud boxes\n`,
   };
-}
-
-/**
- * Where to run `gh`, and whether it must be told the repo.
- *
- * `gh` infers the repo from its cwd's git remote, so the host checkout is the
- * natural place to run it. A control box has no checkout: passing that path as
- * `cwd` makes the spawn itself fail with a bare `spawn gh ENOENT` (Node reports
- * a missing cwd exactly like a missing binary), which reads as "gh isn't
- * installed" and sent us hunting the wrong problem. Fall back to a directory
- * that exists and name the repo explicitly from the registered origin instead.
- */
-export function ghRunContext(
-  workspacePath: string,
-  originUrl: string | undefined,
-  args: string[],
-): { cwd: string; args: string[] } {
-  if (workspacePath.length > 0 && existsSync(workspacePath)) {
-    return { cwd: workspacePath, args };
-  }
-  const origin = originUrl?.trim() ?? '';
-  const alreadyScoped = args.some((a) => a === '--repo' || a === '-R' || a.startsWith('--repo='));
-  if (origin.length === 0 || alreadyScoped) return { cwd: tmpdir(), args };
-  try {
-    return { cwd: tmpdir(), args: ['--repo', repoSlugFromRemote(origin), ...args] };
-  } catch {
-    return { cwd: tmpdir(), args };
-  }
 }
 
 /**
